[SPT-818] rename storage constants file
